# server/docs/swagger.yaml
basePath: /
definitions:
...
  database.RewatchRecord:
    properties:
      created_at:
        type: string
      record_id:
        description: record id for movie/show
        type: integer
      rewatch_finished_at:
        type: string
      rewatch_id:
        type: integer
      rewatch_started_at:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  database.WatchActivity:
    properties:
      created_at:
        type: string
      duration:
        description: duration/runtime in minutes
        type: integer
      episode_number:
        type: integer
      media_source:
        description: tmdb, openlibrary, etc. the main metadata provider
        type: string
      media_title:
        description: movie, tvshow, season or episode title
        type: string
      overview:
        description: game of thrones is a show about ...
        type: string
      record_id:
        type: integer
      record_type:
        description: movie,episode
        type: string
      release_date:
        description: 2012-12-30, for shows/seasons - first_air_date, for episodes
          - air_date
        type: string
      rewatch_id:
        type: integer
      season_number:
        type: integer
      show_media_title:
        type: string
      show_source_id:
        type: string
      source_id:
        description: tmdb id, episode/movie tmdb id
        type: string
      updated_at:
        type: string
      watch_event_id:
        type: integer
      watch_type:
        description: watch, scrobble
        type: string
      watched_at:
        type: string
    type: object
  database.WatchEventMediaRecord:
    properties:
      created_at:
        type: string
      duration:
        description: duration/runtime in minutes
        type: integer
      episode_number:
        type: integer
      media_source:
        description: tmdb, openlibrary, etc. the main metadata provider
        type: string
      media_title:
        description: movie, tvshow, season or episode title
        type: string
      overview:
        description: game of thrones is a show about ...
        type: string
      record_id:
        type: integer
      record_type:
        description: movie,episode
        type: string
      release_date:
        description: 2012-12-30, for shows/seasons - first_air_date, for episodes
          - air_date
        type: string
      rewatch_id:
        type: integer
      season_number:
        type: integer
      source_id:
        description: tmdb id, episode/movie tmdb id
        type: string
      updated_at:
        type: string
      watch_event_id:
        type: integer
      watch_type:
        description: watch, scrobble
        type: string
      watched_at:
        type: string
    type: object
...
  model.WatchHistoryMoviePayload:
    properties:
      action_type:
        enum:
        - watch
        - scrobble
        type: string
      watched_at:
        type: string
    required:
    - action_type
    type: object
  model.WatchHistoryTVShowPayload:
    properties:
      action_type:
        enum:
        - watch
        - scrobble
        type: string
      episode_ids:
        description: tmdb unique id for episode
        items:
          type: integer
        type: array
      episode_number:
        type: integer
      rewatch_id:
        type: integer
      season_number:
        type: integer
      watched_at:
        type: string
    required:
    - action_type
    type: object
...
  v1.AddWatchHistoryMovieResponse:
    properties:
      action_type:
        type: string
      inserted_source_id:
        type: integer
      media_source:
        type: string
    type: object
  v1.AddWatchHistoryTVResponse:
    properties:
      inserted_episode_ids:
        items:
          type: integer
        type: array
      media_source:
        type: string
      skipped_episode_ids:
        items:
          type: integer
        type: array
    type: object
...
  v1.DeleteWatchHistoryPayload:
    properties:
      watch_event_ids:
        items:
          type: integer
        type: array
    required:
    - watch_event_ids
    type: object
...
  v1.V1ErrorResponse:
    properties:
      error:
        example: error message
        type: string
      status:
        enum:
        - success
        - error
        example: error
        type: string
    type: object
  v1.V1SuccessResponse:
    properties:
      data: {}
      status:
        enum:
        - success
        - error
        example: success
        type: string
    type: object
...
  view.MediaRewatchRecordWatchEvents:
    properties:
      created_at:
        type: string
      record_id:
        description: record id for movie/show
        type: integer
      rewatch_finished_at:
        type: string
      rewatch_id:
        type: integer
      rewatch_started_at:
        type: string
      target_season:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      watch_events:
        items:
          $ref: '#/definitions/database.WatchEventMediaRecord'
        type: array
    type: object
...
  view.WatchActivityResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total_records:
        type: integer
      watch_activity:
        items:
          $ref: '#/definitions/database.WatchActivity'
        type: array
    type: object
host: localhost:2323
info:
  contact: {}
  description: |-
    V1 Hound Media Server API Documentation
    While in beta, please expect breaking api changes in short/no notice
  title: Hound API Documentation V1
  version: beta-0.1.0
paths:
...
  /movie/{id}/history/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Watch Event IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.DeleteWatchHistoryPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Delete Movie Watch History
      tags:
      - Watch History
  /tv/{id}/history/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Watch Event IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.DeleteWatchHistoryPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Delete TV Show Watch History
      tags:
      - Watch History
...
  /v1/movie/{id}/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.MediaRewatchRecordWatchEvents'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get Movie Watch History
      tags:
      - Watch History
    post:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Watch History Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.WatchHistoryMoviePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.AddWatchHistoryMovieResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Add Movie Watch History
      tags:
      - Watch History
...
  /v1/tv/{id}/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Season number
        in: query
        name: seasonNumber
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.MediaRewatchRecordWatchEvents'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get TV Show Watch History
      tags:
      - Watch History
    post:
      consumes:
      - application/json
      description: |-
        Only one of season_number and episode_number, or []episode_ids should be defined.
        By default, if []episode_ids are used, this won't clear existing playback progress for
        those episodes, however, this behavior might change in the future. For the season_number,
        episode_number case, the current playback progress will be cleared if the watched_at is newer
        than the existing playback progress.
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Watch History Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.WatchHistoryTVShowPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.AddWatchHistoryTVResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Add TV Show Watch History
      tags:
      - Watch History
  /v1/tv/{id}/history/rewatch:
    post:
      consumes:
      - application/json
      description: Create new rewatch for tv show. This archives the previous watches,
        so user's can start fresh.
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/database.RewatchRecord'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Create TV Show Rewatch
      tags:
      - Watch History
...
  /v1/tv/{id}/season/{seasonNumber}/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Season number
        in: query
        name: seasonNumber
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.MediaRewatchRecordWatchEvents'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get TV Show Watch History
      tags:
      - Watch History
...
  /v1/watch_activity:
    get:
      consumes:
      - application/json
      parameters:
      - description: Limit - Defaults at 500
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Start time in RFC3339 format
        example: "2026-03-13T10:20:30Z"
        in: query
        name: start_time
        type: string
      - description: End time in RFC3339 format
        example: "2026-03-13T10:20:30Z"
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/view.WatchActivityResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get Watch Activity for User
      tags:
      - Watch Activity
swagger: "2.0"